# management _sysytem/app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App Name API Documentation",
...
        "version": "1.0.0",
        "x-logo": {
            "url": "http://laravel-project-template.test/logo.png",
            "backgroundColor": "#FFFFFF",
            "altText": "LDE"
        }
    },
    "servers": [
        {
            "url": "http://laravel-project-template.test",
            "description": "Production Server"
        },
        {
            "url": "http://laravel-project-template.test",
            "description": "Staging Server"
        },
        {
            "url": "http://laravel-project-template.test/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/candidacy_histories/{candidacy}": {
            "get": {
                "tags": [
                    "CandidacyHistory"
                ],
                "summary": "Get All histories data of a candidacy",
                "description": "Returns all history data of a candidacy",
                "operationId": "getListOfHistoriesForACandidacy",
                "parameters": [
                    {
                        "name": "candidacy",
                        "in": "path",
                        "description": "Candidacy key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Histories retrieved successfully"
                                        },
                                        "data": {
                                            "description": "Array of data of Histories",
                                            "properties": {
                                                "histories": {
                                                    "description": "Get histories of a candidacy",
                                                    "items": {
                                                        "properties": {
                                                            "key": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/key"
                                                            },
                                                            "actor": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/actor"
                                                            },
                                                            "stage_name": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/stage_name"
                                                            },
                                                            "status": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/status"
                                                            },
                                                            "metadata": {
                                                                "description": "Get histories of a candidacy",
                                                                "type": "object",
                                                                "example": {
                                                                    "duration": "2hour",
                                                                    "interviewDate": "2021-03-24"
                                                                }
                                                            },
                                                            "created_at": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/created_at"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getListOfUsers",
                "responses": {
                    "200": {
                        "description": "Successfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "description": "Array of data of all Users",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/candidacy_histories/{candidacy}": {
            "get": {
                "tags": [
                    "AdminCandidacyHistory"
                ],
                "summary": "Get All histories data of a candidacy",
                "description": "Returns all history data of a candidacy",
                "operationId": "getListOfHistoriesForACandidacy",
                "parameters": [
                    {
                        "name": "candidacy",
                        "in": "path",
                        "description": "Candidacy key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucessfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Histories retrieved successfully"
                                        },
                                        "data": {
                                            "description": "Array of data of Histories",
                                            "properties": {
                                                "histories": {
                                                    "description": "Get histories of a candidacy",
                                                    "items": {
                                                        "properties": {
                                                            "key": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/key"
                                                            },
                                                            "actor": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/actor"
                                                            },
                                                            "stage_name": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/stage_name"
                                                            },
                                                            "status": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/status"
                                                            },
                                                            "metadata": {
                                                                "description": "Get histories of a candidacy",
                                                                "type": "object",
                                                                "example": {
                                                                    "duration": "2hour",
                                                                    "interviewDate": "2021-03-24"
                                                                }
                                                            },
                                                            "created_at": {
                                                                "$ref": "#/components/schemas/CandidacyHistory/properties/created_at"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "response5xx": {
                "description": "Response for 5xx type of errors",
                "properties": {
                    "success": {
                        "description": "Shows status of operation",
                        "type": "boolean",
                        "default": false
                    },
                    "message": {
                        "description": "Error message to be shown to user",
                        "type": "string",
                        "example": "Database error."
                    },
                    "error": {
                        "description": "Full error for debugging",
                        "example": {
                            "key1": "data1",
                            "key2": "data2"
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User Schema",
                "description": "Schema of User Response",
                "properties": {
                    "key": {
                        "description": "User key to access it",
                        "type": "string",
                        "example": "RLvAbH"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "Allen Walsh"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "werdman@example.org"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "Header": {
                "type": "apiKey",
                "description": "This API uses TOKEN KEY authentication\n\nKey MUST be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "header",
                "scheme": "https"
            },
            "Token": {
                "type": "apiKey",
                "description": "This API uses **TOKEN KEY** authentication\n\nKey `MUST` be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "query",
                "scheme": "https"
            }
        }
    },
...
}